diagnosis/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.10
django-crontab==0.7.1
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
ndg-httpsclient==0.5.1
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.1
psycopg2-binary==2.8.6
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
semantic-version==2.8.5
setuptools-rust==0.11.6
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
Unipath==1.1
urllib3==1.26.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
